backend/pyproject.toml
[tool.poetry]
name = "blind-chess-backend"
version = "0.1.0"
description = "Backend for blind chess"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0"
asyncpg = "^0.27.0"
alembic = "^1.12.0"
python-chess = "^1.999"
python-socketio = "^5.9"
aioredis = "^2.0"
pydantic = "^1.10"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
